requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
aiosqlite==0.19.0
httpx==0.26.0
jinja2==3.1.3
python-multipart==0.0.6
openai==1.10.0
anthropic==0.18.0
pillow==10.2.0
replicate==0.22.0
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0


